Update to v1.1
= 1.1 =
* Better checking to make sure the user id is valid has now been
implemented. If not valid then it uses data from the current user.
* Most functions now use better and more accurate functions to grab
information
* Image and url functions now work better and have methods of showing
the url in a customizable method
* New parameters for the profile image shortcode such as height, width,
and show = yes to show the image as html instead of just the link.
* New parameters for the profile url such as show = yes that will show
the link instead of just the url, before = and after = to add text or
html before or after whatever is outputted, and profile_page = to set
which page of the profile to provide the link for
* More functions to come!